requirements.txt
requests
gspread
oauth2client
pillow
mwclient